Almost working POC for editing components on an entity.
Manually constructing the Inspector panel one component at a time, wired-up to upsert any given component when one of its values changes.
Updated the Ageny flame component to use `update(dt)` to set its position (and MoveToEffect) so changing its `LocalPosition`'s x/y values in the Inspector panel will be immediately reflected on the game screen.

Tbd how to do the reverse. Eg, clicking on a mob that moves, and seeing the LocalPosition values in the Inspector get updated.